requirements.txt
langchain==1.2.0
langchain-openai==0.3.1
pandas==1.4.0
requests==2.26.0
openai==0.27.0
streamlit
streamlit-pills
streamlit-aggrid==0.2.2-2
hydralit
python-dotenv